fbproject/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.52
pydantic==1.10.12
kafka-python==2.0.2
requests==2.31.0
